no more refresh_token_logic+MINOR changes
