+++ service/api_logic/models/api_models.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+REFRESH = "refresh"
 
 class AuthStrategies(Enum):
     SIMPLE = "simple"
