+++ service/api_logic/user_logic.py
@@ -1,4 +1,5 @@
-from flask import current_app, url_for, jsonify, make_response
+from flask import current_app, url_for, jsonify, make_response, request
+from user_agents import parse
 from flask_mail import Message
 from itsdangerous import URLSafeTimedSerializer, SignatureExpired, BadSignature
 from dto.api_input import UpdateUserPreferencesDTO
@@ -20,7 +21,11 @@ from database.postgres.dto.jwt import JwtDTO
 from database.postgres.dto.refresh import RefreshTokenDTO
 from database.postgres.dto.additional_claims import AdditionalClaimsDTO
 from database.postgres.dto.responce_data import ResponceDataDTO
+from database.postgres.dto.device_info import DeviceInfoDTO
 from datetime import datetime
+import time
+import requests
+import hashlib
 from flask_jwt_extended import get_jwt_identity, get_jwt
 from service.api_logic.models.api_models import SportPreferenceFields, TeamPreferenceFields
 
@@ -31,15 +36,70 @@ TEAM_TYPE = "team"
 
 class UserService:
 
-    def __init__(self, user_dal, preferences_dal, sport_dal, access_tokens_dal, refresh_dal, user_info_service):
+    def __init__(self, user_dal, preferences_dal, sport_dal, access_tokens_dal, refresh_dal):
         self._user_dal = user_dal
         self._access_tokens_dal = access_tokens_dal
         self._refresh_dal = refresh_dal
         self._preferences_dal = preferences_dal
         self._sport_dal = sport_dal
-        self._user_info_service = user_info_service
         self._serializer = URLSafeTimedSerializer(current_app.secret_key)
         self._logger = Logger("logger", "all.log").logger
+    @staticmethod    
+    def get_user_device(self) -> str:
+        user_agent = request.headers.get("User-Agent", "")
+        parsed_agent = parse(user_agent)
+
+        device_info = DeviceInfoDTO(
+            browser=f"{parsed_agent.browser.family} {parsed_agent.browser.version_string}",
+            os=f"{parsed_agent.os.family} {parsed_agent.os.version_string}",
+            device=parsed_agent.device.family
+        )
+        
+        return device_info
+
+    def __get_client_ip(self) -> str:
+        return request.headers.get("X-Forwarded-For", request.remote_addr)
+
+    def __get_country_from_ip(self) -> str:
+        ip = self.__get_client_ip()
+
+        try:
+            response = requests.get(f"https://ipinfo.io/{ip}/json", timeout=3)
+            response.raise_for_status()
+            unknown = "Unknown"
+            return response.json().get("country", unknown)
+        except (requests.RequestException, ValueError):
+            return unknown
+
+    def has_ip_country_changed(self, stored_country: str) -> bool:
+        current_country = self.get_country_from_ip()
+        return stored_country != current_country
+
+    def has_device_changed(self, stored_device: str) -> bool:
+        current_device = self.get_user_device()
+        return stored_device != current_device
+
+    def generate_nonce(self):
+        nonce = hashlib.sha256(f"{time.time()}{os.urandom(16)}".encode()).hexdigest()
+        return nonce
+    
+    def is_suspicious_login(self, user_id: int) -> bool:
+        refresh_entry = self._access_token_dal.get_valid_refresh_token_by_user(user_id)
+        if not refresh_entry:
+            return False  
+
+        current_ip = self.__get_client_ip()
+        current_country = self.__get_country_from_ip()
+        current_device = self.get_user_device()
+
+        suspicious_conditions = [
+            refresh_entry.last_ip and refresh_entry.last_ip != current_ip,
+            refresh_entry.last_country and refresh_entry.last_country != current_country,
+            refresh_entry.last_device and refresh_entry.last_device != current_device,
+            self._access_token_dal.is_nonce_used(user_id, refresh_entry.nonce)
+        ]
+
+        return any(suspicious_conditions)
 
 
     def get_user_by_email_or_username(self, email=None, username=None):
@@ -145,7 +205,7 @@ class UserService:
     async def log_in(self, credentials: InputUserLogInDTO):
         login_context = AuthManager(self)
         user_id = self._user_dal.get_existing_user(credentials.email)
-        sus_login = self._user_info_service.is_suspicious_login(user_id)
+        sus_login = self.is_suspicious_login(user_id)
         if not sus_login:
             user = await login_context.execute_log_in(credentials)
             response = await self.create_access_token_response(user)
@@ -194,9 +254,9 @@ class UserService:
 
         refresh_dto = RefreshTokenDTO(
             user_id=user.id,
-            last_ip=self._user_info_service.get_country_from_ip(),
-            last_device=self._user_info_service.get_user_device(),
-            nonce=self._user_info_service.generate_nonce()
+            last_ip=self.get_country_from_ip(),
+            last_device=self.get_user_device(),
+            nonce=self.generate_nonce()
         )
         self._refresh_dal.save_refresh_token(refresh_dto)
 
@@ -266,7 +326,7 @@ class UserService:
         )
         new_access_token = create_access_token(identity=email.email, additional_claims=additional_claims)
         new_refresh_token = create_refresh_token(identity=email.email, 
-                                                 additional_claims=additional_claims.update({"nonce": self._user_info_service.generate_nonce()}))
+                                                 additional_claims=additional_claims.update({"nonce": self.generate_nonce()}))
         
         return new_access_token, new_refresh_token
     
@@ -276,10 +336,10 @@ class UserService:
         
         refresh_dto = RefreshTokenDTO(
             user_id=user.id,
-            last_ip=self._user_info_service.get_country_from_ip(),
-            last_device=self._user_info_service.get_user_device(),
+            last_ip=self.__get_country_from_ip(),
+            last_device=self.get_user_device(),
             refresh_token=new_refresh_token,
-            nonce=self._user_info_service.generate_nonce()
+            nonce=self.generate_nonce()
         )
 
         self._refresh_dal.update_refresh_token(user.id, refresh_dto)
@@ -295,7 +355,7 @@ class UserService:
 
         new_access_token, new_refresh_token = await self.create_new_access_and_refresh_tokens(identity, refresh=True)
 
-        new_nonce = self._user_info_service.generate_nonce()
+        new_nonce = self.generate_nonce()
         self._refresh_dal.update_refresh_token(identity, new_refresh_token, new_nonce)
 
         response = make_response(jsonify({
